Issues at corners, front wall mostly fixed
I'm still unhappy with the collisions around some of the corners, but I need to move on to applying rotation to the hit-boxes for now.  Once rotation is solved, I can apply collisions to the other objects, work on fixing the lighting on Cuboid and Cabin objects then come back to the corner issues if time permits.