Lantern collisions.
All objects have basic collisions functional, now I need to finalize the area design, rework the lighting and user controls, and redo the dimensions and normals of the Cabin object.  If time permits, I want to spend more time on the skybox, what I have is technically sufficient, but I'm not quite happy with how it looks currently.